3D Model Editor: -Added options to the edit panel for rotating the UV of a given face by 90 degrees. -Updated triangle render.
Engine:
-Updated triangle render.
-Added a world floor.
-New lighting mechanics for light sources: After 1 second of not detecting the light source that created the lightmap, the lightmap is removed. If the lightmap is unlit for 1 second, it is also removed. For objects larger than [largeFaceThreshold], instead of creating one large lightmap, lightmaps of dimensions [cubeSize] x [cubeSize] are dynamically created.

Updated triangle render.

The renderer takes the positions of the triangle vertices and uses them to create a 4th vertex, so that instead of a triangle, it would technically be a square. Based on the square, it simply applies a square texture to our imaginary square. The texture is applied as with any other surface. Then, it cuts out an additional section after the 4th vertex. And our old triangle area remains – this is what is rendered in the final rendering.

Simplification: Convert triangle to square -> Apply texture to square -> Return to triangle dimensions.